+++ src/rgxlog-interpreter/src/rgxlog/engine/state/term_graph.py
